refactor:salaryDetailSection 반복되는 구조 컴포넌트화
